索引操作和rename并发时，捕获索引操作加锁失败异常

+++ testcase_new/story/java/src/test/java/com/sequoiadb/index/serial/IndexConsistent23958.java
@@ -80,6 +80,8 @@ public class IndexConsistent23958 extends SdbTestBase {
             if ( dropIndex.getRetCode() != SDBError.SDB_DMS_CS_NOTEXIST
                     .getErrorCode()
                     && dropIndex.getRetCode() != SDBError.SDB_DMS_NOTEXIST
+                            .getErrorCode()
+                    && dropIndex.getRetCode() != SDBError.SDB_LOCK_FAILED
                             .getErrorCode() ) {
                 Assert.fail( "---drop index fail!" + dropIndex.getRetCode() );
             }
